// pslpwcl/src/app/modules/pslpapi/model/schedaAnagraficaProfessionale.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AccorpTransazione } from './accorpTransazione';
import { DatiAmministrativi } from './datiAmministrativi';
import { DatiAnagrafici } from './datiAnagrafici';
import { Decodifica } from './decodifica';
import { EsperienzeLavorative } from './esperienzeLavorative';
import { InformazioniCurriculari } from './informazioniCurriculari';
import { PoliticheAttive } from './politicheAttive';

export interface SchedaAnagraficaProfessionale { 
    accorpTransazione?: Array<AccorpTransazione>;
    dataUltimoAggiornamento?: Date;
    datiAmministrativi?: DatiAmministrativi;
    datiAnagrafici?: DatiAnagrafici;
    enteTitolare?: Decodifica;
    esperienzeLavorative?: EsperienzeLavorative;
    fonte?: Decodifica;
    idInterscambioSap?: number;
    idSilLavAnagrafica?: number;
    identificativoSap?: string;
    informazioniCurriculari?: InformazioniCurriculari;
    politicheAttiveLavoratore?: PoliticheAttive;
}