// pslpwcl/src/app/modules/pslpapi/model/altroCorso.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Decodifica } from './decodifica';
import { LavCorsoForm } from './lavCorsoForm';
import { LavCorsoFormAltro } from './lavCorsoFormAltro';

export interface AltroCorso { 
    ambitoDisciplinare?: Decodifica;
    annoGestione?: string;
    codice?: string;
    denominazioneCorso?: string;
    flgCv?: boolean;
    idFormazioneProfessionale?: number;
    lavCorsoForm?: LavCorsoForm;
    lavCorsoFormAltro?: LavCorsoFormAltro;
    numDurataOre?: number;
}