// pslpwcl/src/app/modules/pslpapi/model/corsoFormazioneRequest.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AltroCorsoForm } from './altroCorsoForm';
import { CallInfo } from './callInfo';
import { LavCorsoForm } from './lavCorsoForm';
import { LavCorsoFormAltro } from './lavCorsoFormAltro';

export interface CorsoFormazioneRequest { 
    callInfo?: CallInfo;
    lavCorsoForm?: LavCorsoForm;
    lavCorsoFormAltro?: LavCorsoFormAltro;
    silTAltroCorsoForm?: AltroCorsoForm;
}